// data/github/current/ui/packages/issue-metadata/hooks/__generated__/useIssueTypeSuggestionsQuery.graphql.ts
export type useIssueTypeSuggestionsQuery$variables = {
  issueId: string;
};
export type useIssueTypeSuggestionsQuery$data = {
  readonly node: {
    readonly body?: string;
    readonly repository?: {
      readonly issueTypes: {
        readonly nodes: ReadonlyArray<{
          readonly " $fragmentSpreads": FragmentRefs<"IssueTypePickerIssueType">;
        } | null | undefined> | null | undefined;
      } | null | undefined;
    };
    readonly title?: string;
  } | null | undefined;
};
export type useIssueTypeSuggestionsQuery = {
  response: useIssueTypeSuggestionsQuery$data;
  variables: useIssueTypeSuggestionsQuery$variables;
};